Clean up some syscall logic

+++ userspace/library/include/syscalls.h
@@ -5,8 +5,8 @@ void write_str(char* string);
 void delay_us(unsigned long microseconds);
 int get_pid();
 int open(char* path);
-int read(int index, char* buffer, int length);
-int close(int index);
+unsigned long read(unsigned long index, char* buffer, unsigned long length);
+int close(unsigned long index);
 
 
 #endif //BEKOS_LIBC_SYSCALLS_H
